requirements.txt
flask
torch
transformers
Pillow
requests
spotipy
